# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MédiGuide API",
    "description": "API de localisation de professionnels de santé (médecins, pharmacies, parapharmacies) à Jendouba, Tunisie : recherche géographique, prise de rendez-vous, avis, espace professionnel et administration. Un module complémentaire (medieye) expose ces mêmes fiches au format Facility générique de la plateforme partagée mediEYE/craftEYE/cityEYE sous /api/modules/medieye — voir openapi-medieye-fragment.json.",
    "version": "1.0.0",
    "contact": {
      "name": "MédiGuide"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Développement local"
    },
    {
      "url": "https://<votre-projet>.up.railway.app/api",
      "description": "Production (Railway)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token",
        "description": "JWT d'accès posé en cookie httpOnly après /api/auth/login"
      },
      "csrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "Requis sur toute requête POST/PATCH/DELETE ; valeur du cookie csrf_token obtenu via GET /api/auth/csrf"
      }
    },
    "schemas": {
...
      "Facility": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "nom": {
            "type": "string"
          },
          "category": {
            "type": "string",
            "enum": [
              "medecin",
              "pharmacie",
              "parapharmacie"
            ]
          },
          "specialite": {
            "type": "string",
            "nullable": true
          },
          "adresse": {
            "type": "string"
          },
          "delegation": {
            "type": "string"
          },
          "telephone": {
            "type": "string",
            "nullable": true
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          },
          "noteAvis": {
            "type": "number",
            "format": "float",
            "nullable": true
          },
          "bio": {
            "type": "string",
            "nullable": true
          },
          "photoUrl": {
            "type": "string",
            "nullable": true
          },
          "distanceMetres": {
            "type": "number",
            "nullable": true,
            "description": "Présent uniquement si lat/lng fournis dans la recherche"
          }
        }
      },
...
      "RendezVous": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "disponibiliteId": {
            "type": "integer"
          },
          "patientId": {
            "type": "integer"
          },
          "statut": {
            "type": "string",
            "enum": [
              "confirme",
              "annule"
            ]
          },
          "motif": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    }
  },
  "security": [
    {
      "cookieAuth": []
    }
  ],
  "paths": {
...
    "/facilities": {
      "get": {
        "tags": [
          "Facilities"
        ],
        "summary": "Rechercher médecins / pharmacies / parapharmacies",
        "security": [],
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "medecin",
                "pharmacie",
                "parapharmacie"
              ]
            }
          },
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Recherche full-text nom + spécialité"
          },
          {
            "name": "specialite",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ville",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "typeGarde",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "lat",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "lng",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "radius",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 5000
            },
            "description": "Rayon en mètres, plafonné à 50000"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 300
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des établissements (résultat mis en cache Redis 60s si géo)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Facility"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Catégorie invalide"
          }
        }
      }
    },
...
    "/facilities/{id}": {
      "get": {
        "tags": [
          "Facilities"
        ],
        "summary": "Détail d'un établissement",
        "security": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Fiche détaillée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Facility"
                }
              }
            }
          },
          "404": {
            "description": "Introuvable"
          }
        }
      }
    },
...
    "/rendezvous": {
      "post": {
        "tags": [
          "Rendez-vous"
        ],
        "summary": "Réserver un créneau",
        "description": "Transaction avec verrou de ligne (SELECT ... FOR UPDATE) sur la disponibilité pour empêcher une double réservation ; envoie un e-mail de confirmation.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "disponibiliteId"
                ],
                "properties": {
                  "disponibiliteId": {
                    "type": "integer"
                  },
                  "motif": {
                    "type": "string",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Rendez-vous confirmé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RendezVous"
                }
              }
            }
          },
          "403": {
            "description": "Rôle patient requis"
          },
          "404": {
            "description": "Créneau introuvable"
          },
          "409": {
            "description": "Créneau déjà réservé entre-temps"
          }
        }
      }
    },
...
    "/rendezvous/{id}/annuler": {
      "patch": {
        "tags": [
          "Rendez-vous"
        ],
        "summary": "Annuler un rendez-vous",
        "description": "Libère le créneau associé (estReserve = false).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Annulé"
          },
          "403": {
            "description": "Auteur uniquement"
          }
        }
      }
    },
    "/rendezvous/aujourdhui": {
      "get": {
        "tags": [
          "Rendez-vous"
        ],
        "summary": "Rendez-vous confirmés du jour (professionnel connecté)",
        "responses": {
          "200": {
            "description": "Rendez-vous du jour"
          },
          "403": {
            "description": "Rôle professionnel requis"
          }
        }
      }
    },
...
  }
}